+++ semantic-conventions/src/tests/semconv/templating/test_markdown.py
@@ -128,7 +128,7 @@ class TestCorrectMarkdown(unittest.TestCase):
 
     def test_scope(self):
         self.check("markdown/scope/")
-    
+
     def test_attribute_group(self):
         self.check("markdown/attribute_group/")