+++ semantic-conventions/CHANGELOG.md
@@ -5,7 +5,7 @@ Please update the changelog as part of any significant pull request.
 ## Unreleased
 
 - Add a semantic convention type for generic attribute group ("attribute_group")
-  ([#TODO](https://github.com/open-telemetry/build-tools/pull/TODO)).
+  ([#124](https://github.com/open-telemetry/build-tools/pull/124)).
 
 ## v0.14.0