+++ semantic-conventions/CHANGELOG.md
@@ -6,6 +6,8 @@ Please update the changelog as part of any significant pull request.
 
 - Add a semantic convention type for Metrics ("metric" and "metric_group") 
   ([#79](https://github.com/open-telemetry/build-tools/pull/79))
+- Add a semantic convention type for generic attribute group ("attribute_group")
+  ([#124](https://github.com/open-telemetry/build-tools/pull/124)).
 
 ## v0.14.0